Rename from api to network folder.
