+++ app/src/main/java/ru/sandbox/androidacademyapp/MoviesViewModelFactory.kt
@@ -3,7 +3,7 @@ package ru.sandbox.androidacademyapp
 import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import ru.sandbox.androidacademyapp.data.api.MoviesApi
+import ru.sandbox.androidacademyapp.data.network.MoviesApi
 import ru.sandbox.androidacademyapp.data.db.MovieDatabase
 import ru.sandbox.androidacademyapp.repository.MovieRepository
