+++ app/src/main/java/ru/sandbox/androidacademyapp/data/network/GenreResponse.kt
@@ -1,4 +1,4 @@
-package ru.sandbox.androidacademyapp.data.api
+package ru.sandbox.androidacademyapp.data.network
 
 import kotlinx.serialization.Serializable
