+++ app/src/main/java/ru/sandbox/androidacademyapp/data/network/MoviesApiHeaderInterceptor.kt
@@ -1,4 +1,4 @@
-package ru.sandbox.androidacademyapp.data.api
+package ru.sandbox.androidacademyapp.data.network
 
 import okhttp3.Interceptor
 import okhttp3.Response
