+++ app/src/main/java/ru/sandbox/androidacademyapp/data/network/Result.kt
@@ -1,4 +1,4 @@
-package ru.sandbox.androidacademyapp.data.api
+package ru.sandbox.androidacademyapp.data.network
 
 sealed class Result<T>(
     val data: T? = null,
